feat: implemented code for CadastrarActivity controller. Developed method to validate data before save

+++ app/src/main/java/dahlke/br/fintechapp/FinancialOperation.kt
@@ -6,6 +6,6 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 data class FinancialOperation (
     val description : String,
-    val operationValue : Float,
+    val operationValue : Double,
     val operationType : FinancialOperationType) : Parcelable {
 }
